requirements.txt
Flask
requests
python-dotenv
gunicorn
google-generativeai>=0.3.0
requests>=2.25.1